backend/requirements.txt
# Python 3.13 Compatible Requirements
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database - Using newer versions for Python 3.13 support
sqlalchemy>=2.0.35  # Python 3.13 compatible
greenlet>=3.0.0  # Required for SQLAlchemy async operations
alembic==1.13.1
psycopg[binary]>=3.2.0  # Newer psycopg3 instead of psycopg2-binary
asyncpg==0.30.0  # Latest version with better Python 3.13 support

# Vector Store
pgvector==0.2.5

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic>=2.10.0  # Latest version with Python 3.13 support
pydantic-settings>=2.7.0
email-validator==2.1.0

# AI/ML — LLM Providers (all optional, install what you use)
# AWS Bedrock (optional — uses your AWS credentials, no extra API keys)
boto3>=1.34.34
botocore>=1.34.34

# OpenAI (optional, if using OpenAI directly)
openai>=1.54.0  # Latest with native async support

# Anthropic (optional, if using Anthropic directly)
anthropic==0.18.1

# Data Processing
openpyxl==3.1.2
numpy==1.26.4
pandas==2.2.0
python-docx==1.1.2  # For .doc/.docx file parsing
PyPDF2==3.0.1  # For PDF text extraction

# HTTP Client & Web Scraping
httpx==0.26.0
aiofiles==23.2.1
beautifulsoup4==4.12.3

# Utilities
python-slugify==8.0.3
pydantic-extra-types>=2.10.0

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0

# Background Tasks & Scheduling
APScheduler==3.10.4

# Celery - Durable task queue
celery[redis]==5.3.6
redis==5.0.1

# Instructor - Structured LLM outputs
instructor==1.6.4

# Optional Dependencies:
# - Heavy ML dependencies (local embeddings, clustering): pip install -r requirements-ml.txt
# - Development tools (testing, linting): pip install -r requirements-dev.txt
